Move getNames to a separated file.
